+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 3.1.0
 
+- useDrag
+
 - Reshaped global reset styles in rs.reset
 - Switched to :focus-visible wherever possible
 - Modal: attribures.ref support
@@ -21,7 +23,9 @@
 - Popover, Tooltip: Only trigger focus events for keyboard mode
 - Popover, DropdownMenu: Popover.Dismissible
 - Slider: Fixed focus state to work based on focus-visible
+- Slider: renderValue=false
 
+- Figma: Autocomplete as component
 - Figma: Removed local color scoping
 - Figma: Updated description links
 - Figma: typography weight and scoping